client/requirements-windows.txt
pycaw
pyserial
winrt-Windows.Foundation
winrt-Windows.Foundation.Collections
winrt-Windows.Media
winrt-Windows.Media.Control
winrt-Windows.Storage
winrt-Windows.Storage.Streams